pyproject.toml
[tool.poetry]
name = "nonebot_plugin_imagetools"
version = "0.1.0"
description = "Nonebot2 简单图片操作插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-imagetools"
repository = "https://github.com/noneplugin/nonebot-plugin-imagetools"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
httpx = ">=0.19.0"
imageio = "^2.12.0"
Pillow = "^9.0.0"
nonebot-plugin-imageutils = "^0.1.3"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
